Dashboard: extract SilentErrorBoundary and use it for guided-tour popups
